Move Freeze.kt from Player into Character effect package
