+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/effect/Freeze.kts
@@ -1,4 +1,4 @@
-package world.gregs.voidps.world.interact.entity.player.effect
+package world.gregs.voidps.world.interact.entity.effect
 
 import world.gregs.voidps.engine.client.variable.start
 import world.gregs.voidps.engine.client.variable.stop
