+++ game/src/test/kotlin/world/gregs/voidps/world/interact/entity/player/combat/AncientSpellEffectsTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 import world.gregs.voidps.engine.client.variable.hasClock
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.world.interact.entity.combat.hit.hit
-import world.gregs.voidps.world.interact.entity.player.effect.frozen
+import world.gregs.voidps.world.interact.entity.effect.frozen
 import world.gregs.voidps.world.interact.entity.player.toxin.poisoned
 
 internal class AncientSpellEffectsTest : CombatFormulaTest() {
